requirements.txt
cachetools
certifi
chardet
charset-normalizer
et_xmlfile
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2
idna
oauthlib
openpyxl
pillow
proto-plus
protobuf
pyasn1
pyasn1_modules
pyparsing
python-dotenv
reportlab
requests
requests-oauthlib
rsa
uritemplate
urllib3
